refactor: Update logic in CreatePost to save posts in db and add CreatePost functionality to FeedView
This commit refactors the logic in the CreatePost view to save posts in the database. It also adds the CreatePost functionality to the FeedView, allowing users to create new posts. Additionally, a new function called createPost is created in the PostService to handle the creation of posts. This commit improves the overall functionality and user experience of the app.